Refactor event display and introduce TeamCard widget for improved team representation
